requirements.txt
beautifulsoup4==4.13.4
certifi==2025.7.14
charset-normalizer==3.4.2
dnspython==2.7.0
idna==3.10
pymongo==4.13.2
python-dotenv==1.1.1
requests==2.32.4
schedule==1.2.2
soupsieve==2.7
typing_extensions==4.14.1
urllib3==2.5.0
playwright==1.47.0
python-dotenv==1.0.1
